from the crow's nest
slowly learning about headers…
